Cleanup devtools. Allow for changing alignment.
