+++ src/Avalonia.Diagnostics/Diagnostics/ViewModels/ControlLayoutViewModel.cs
@@ -10,27 +10,58 @@ namespace Avalonia.Diagnostics.ViewModels
     internal class ControlLayoutViewModel : ViewModelBase
     {
         private readonly IVisual _control;
-        private Thickness _marginThickness;
         private Thickness _borderThickness;
-        private Thickness _paddingThickness;
-        private double _width;
         private double _height;
-        private string _widthConstraint;
         private string _heightConstraint;
+        private HorizontalAlignment _horizontalAlignment;
+        private Thickness _marginThickness;
+        private Thickness _paddingThickness;
         private bool _updatingFromControl;
+        private VerticalAlignment _verticalAlignment;
+        private double _width;
+        private string _widthConstraint;
+
+        public ControlLayoutViewModel(IVisual control)
+        {
+            _control = control;
+
+            HasPadding = AvaloniaPropertyRegistry.Instance.IsRegistered(control, Decorator.PaddingProperty);
+            HasBorder = AvaloniaPropertyRegistry.Instance.IsRegistered(control, Border.BorderThicknessProperty);
+
+            if (control is AvaloniaObject ao)
+            {
+                MarginThickness = ao.GetValue(Layoutable.MarginProperty);
+
+                if (HasPadding)
+                {
+                    PaddingThickness = ao.GetValue(Decorator.PaddingProperty);
+                }
+
+                if (HasBorder)
+                {
+                    BorderThickness = ao.GetValue(Border.BorderThicknessProperty);
+                }
+
+                HorizontalAlignment = ao.GetValue(Layoutable.HorizontalAlignmentProperty);
+                VerticalAlignment = ao.GetValue(Layoutable.VerticalAlignmentProperty);
+            }
+
+            UpdateSize();
+            UpdateSizeConstraints();
+        }
 
         public Thickness MarginThickness
         {
             get => _marginThickness;
             set => RaiseAndSetIfChanged(ref _marginThickness, value);
         }
-        
+
         public Thickness BorderThickness
         {
             get => _borderThickness;
             set => RaiseAndSetIfChanged(ref _borderThickness, value);
         }
-        
+
         public Thickness PaddingThickness
         {
             get => _paddingThickness;
@@ -61,35 +92,21 @@ namespace Avalonia.Diagnostics.ViewModels
             private set => RaiseAndSetIfChanged(ref _heightConstraint, value);
         }
 
-        public bool HasPadding { get; }
-        
-        public bool HasBorder { get; }
-        
-        public ControlLayoutViewModel(IVisual control)
+        public HorizontalAlignment HorizontalAlignment
         {
-            _control = control;
-
-            HasPadding = AvaloniaPropertyRegistry.Instance.IsRegistered(control, Decorator.PaddingProperty);
-            HasBorder = AvaloniaPropertyRegistry.Instance.IsRegistered(control, Border.BorderThicknessProperty);
-
-            if (control is AvaloniaObject ao)
-            {
-                MarginThickness = ao.GetValue(Layoutable.MarginProperty);
-
-                if (HasPadding)
-                {
-                    PaddingThickness = ao.GetValue(Decorator.PaddingProperty);
-                }
-
-                if (HasBorder)
-                {
-                    BorderThickness = ao.GetValue(Border.BorderThicknessProperty);
-                }
-            }
-
-            UpdateSize();
-            UpdateSizeConstraints();
+            get => _horizontalAlignment;
+            private set => RaiseAndSetIfChanged(ref _horizontalAlignment, value);
         }
+
+        public VerticalAlignment VerticalAlignment
+        {
+            get => _verticalAlignment;
+            private set => RaiseAndSetIfChanged(ref _verticalAlignment, value);
+        }
+
+        public bool HasPadding { get; }
+
+        public bool HasBorder { get; }
 
         private void UpdateSizeConstraints()
         {
@@ -151,6 +168,14 @@ namespace Avalonia.Diagnostics.ViewModels
                 {
                     ao.SetValue(Border.BorderThicknessProperty, BorderThickness);
                 }
+                else if (e.PropertyName == nameof(HorizontalAlignment))
+                {
+                    ao.SetValue(Layoutable.HorizontalAlignmentProperty, HorizontalAlignment);
+                }
+                else if (e.PropertyName == nameof(VerticalAlignment))
+                {
+                    ao.SetValue(Layoutable.VerticalAlignmentProperty, VerticalAlignment);
+                }
             }
         }
 
@@ -171,21 +196,29 @@ namespace Avalonia.Diagnostics.ViewModels
                         if (e.Property == Layoutable.MarginProperty)
                         {
                             MarginThickness = ao.GetValue(Layoutable.MarginProperty);
-                        } 
+                        }
                         else if (e.Property == Decorator.PaddingProperty)
                         {
                             PaddingThickness = ao.GetValue(Decorator.PaddingProperty);
-                        } 
+                        }
                         else if (e.Property == Border.BorderThicknessProperty)
                         {
                             BorderThickness = ao.GetValue(Border.BorderThicknessProperty);
-                        } 
+                        }
                         else if (e.Property == Layoutable.MinWidthProperty ||
                                  e.Property == Layoutable.MaxWidthProperty ||
                                  e.Property == Layoutable.MinHeightProperty ||
                                  e.Property == Layoutable.MaxHeightProperty)
                         {
                             UpdateSizeConstraints();
+                        }
+                        else if (e.Property == Layoutable.HorizontalAlignmentProperty)
+                        {
+                            HorizontalAlignment = ao.GetValue(Layoutable.HorizontalAlignmentProperty);
+                        }
+                        else if (e.Property == Layoutable.VerticalAlignmentProperty)
+                        {
+                            VerticalAlignment = ao.GetValue(Layoutable.VerticalAlignmentProperty);
                         }
                     }
                 }
